+++ tests/runtime/ray/regression/test_workflow_outputs.py
@@ -6,10 +6,10 @@ from typing import Any, Dict, Tuple, Union
 
 import pytest
 
+from orquestra.sdk._base._config import LOCAL_RUNTIME_CONFIGURATION
 from orquestra.sdk._base._testing import _connections
 from orquestra.sdk._base.serde import deserialize
 from orquestra.sdk._ray import _client, _dag
-from orquestra.sdk.schema import configs
 
 PRODUCING_SDK_VERSIONS_TO_TEST = ["0.46.0", "0.47.0"]
 BASE_PATH = Path(__file__).parent / "data"
@@ -19,18 +19,6 @@ BASE_PATH = Path(__file__).parent / "data"
 pytestmark = pytest.mark.filterwarnings(
     "ignore::pytest.PytestUnraisableExceptionWarning"
 )
-
-
-@pytest.fixture(scope="module")
-def runtime(tmp_path_factory: pytest.TempPathFactory, change_db_location):
-    project_dir = tmp_path_factory.mktemp("ray-regression")
-    config = configs.RuntimeConfiguration(
-        config_name="ray-regression-config",
-        runtime_name=configs.RuntimeName.RAY_LOCAL,
-    )
-    client = _client.RayClient()
-    rt = _dag.RayRuntime(client, config, project_dir)
-    yield rt
 
 
 # Uses real Ray connection
@@ -44,6 +32,14 @@ class TestOutputs:
         storage_path = str(BASE_PATH / request.param)
         with _connections.make_ray_conn(storage_path) as ray_params:
             yield ray_params
+
+    @pytest.fixture(scope="class")
+    def runtime(self, tmp_path_factory: pytest.TempPathFactory, change_db_location):
+        project_dir = tmp_path_factory.mktemp("ray-regression")
+        config = LOCAL_RUNTIME_CONFIGURATION
+        client = _client.RayClient()
+        rt = _dag.RayRuntime(config, project_dir, client)
+        yield rt
 
     @pytest.mark.parametrize(
         "workflow_run_id, expected_result",